Video 26 Completed

+++ app/Http/Controllers/JobController.php
@@ -3,7 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use \App\Mail\JobPosted;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use \Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
@@ -30,23 +35,26 @@ class JobController extends Controller
 			'salary' => ['required']
 		]);
 
-		Job::create([
+		$job = Job::create([
 			'title' => request('title'),
 			'salary' => request('salary'),
 			'employer_id' => 1,
 		]);
+	
+	// Send Job Posted Email
+	Mail::to($job->employer->user)->queue(new JobPosted($job));
 		
 		return (redirect('/jobs'));
 	}
 	
 	public function edit(Job $job)
-	{
+	{		
 		return (view('jobs.edit', ['job' => $job]));
 	}
 	
 	public function update(Job $job)
 	{
-		// Authorize
+		Gate::authorize('edit-job', $job);
 		
 		// validate
 		request()->validate([
@@ -65,7 +73,7 @@ class JobController extends Controller
 	
 	public function destroy(Job $job)
 	{
-		// autorize (On hold...)
+		Gate::authorize('edit-job', $job);
 		
 		// delete the job
 		$job->delete();
